pyproject.toml
[tool.poetry]
name = "notice-board"
version = "0.1.0"
description = ""
authors = ["panthurr0 <olapopaola@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
psycopg = "^3.2.1"
setuptools = "^72.1.0"
black = "^24.8.0"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
